requirements.txt
pillow==10.4.0
djangorestframework==3.15.2
drf-yasg
django-filter==24.3